bisesh branch created

+++ geoapp/views.py
@@ -1,43 +1,57 @@
 from django.shortcuts import render,redirect
 from django.http import HttpResponse
-from .models import Shop,Product
+from .models import *
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import logout,login,authenticate
+from django.contrib import messages
+from django.contrib import messages
+
 # Create your views here.
 @login_required(login_url='login')
 def home(request):
-    shops = Shop.objects.all()
-    return render(request,'home.html',{'shops': shops})
+    if request.user.is_superuser:
+        shops = Shop.objects.all()
+        return render(request,'home.html',{'shops': shops})
+    else:
+        user = request.user 
+        shop = Shop.objects.get(user = user)
+        products = Product.objects.filter(productOwner = shop)
+        return render(request,'home.html',{'products':products})
 
 @login_required(login_url='login')
 def createshop(request):
-    if request.method == 'POST':
-        shopName = request.POST.get('shopName')
-        shopOwner = request.POST.get('shopOwner')
-        shopContact = request.POST.get('shopContact')
-        username = request.POST.get('Username')
-        password = request.POST.get('Password')
-        print(username)
-        # Create a User instance
-        user = User(username=username, password=password)
-        user.save()
-        if user:
-        # Create a Shop instance associated with the created User
-            Shop.objects.create(user=user,  # Link the Shop to the created User
-                                shopName=shopName,
-                                shopOwner=shopOwner,
-                                shopContact=shopContact)
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            shopName = request.POST.get('shopName')
+            shopOwner = request.POST.get('shopOwner')
+            shopContact = request.POST.get('shopContact')
+            username = request.POST.get('Username')
+            password = request.POST.get('Password')
             
-            return redirect('home')  # Redirect to the desired URL after successful creation
-    
-    return render(request, 'createshop.html')
+            if username !=" " and password != "": 
+                user = User(username = username)
+                user.set_password(password)
+                user.save()
+                if user:
+                    createShop = Shop(user = user,shopName = shopName,shopOwner = shopOwner,shopContact = shopContact)
+                    createShop.save()
+            else:
+                messages.success(request,'Fill username or email')
+                
+
+        return render(request, 'createshop.html')
+    else: 
+        return redirect('/')
 
 @login_required(login_url='login')
 def deleteshop(request,pk):
-    shop = Shop.objects.get(shopId = pk)
-    shop.delete()
-    return redirect('home')
+    if request.user.is_superuser:
+        shop = Shop.objects.get(pk = pk)
+        shop.delete()
+        return redirect('home')
+    else:
+        return redirect('/')
 
 @login_required(login_url='login')
 def addproduct(request):
@@ -53,21 +67,37 @@ def addproduct(request):
         return redirect('home')
     return render(request,'addproduct.html')
 
+
+def shopDetail(request,pk): 
+    if request.user.is_superuser:
+        
+        shop = Shop.objects.get(pk = pk)
+        products = Product.objects.filter(productOwner = shop)
+        return render(request,'shopDetails.html',{"shop":shop,'products':products})
+        
+    else:
+        return redirect('/')
+        
 @login_required(login_url='login')
 def editshop(request,pk):
-    shop = Shop.objects.get(shopId = pk)
-    if request.method == 'POST':
-        shop.shopName = request.POST.get('shopName')
-        shop.shopOwner = request.POST.get('shopOwner')
-        shop.shopContact = request.POST.get('shopContact')
-        shop.save()
-        return redirect('home')
-    return render(request,'editshop.html',{'shop':shop})
+    if request.user.is_superuser:
+        shop = Shop.objects.get(pk = pk)
+        if request.method == 'POST':
+            shop.shopName = request.POST.get('shopName')
+            shop.shopOwner = request.POST.get('shopOwner')
+            shop.shopContact = request.POST.get('shopContact')
+            shop.save()
+            return redirect('home')
+        return render(request,'editshop.html',{'shop':shop})
+    else:
+        return redirect('/')
 
 def loginuser(request):
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
+        print(username)
+        print(password)
 
         # Authenticate user
         user = authenticate(request, username=username, password=password)
@@ -86,3 +116,30 @@ def loginuser(request):
 def logoutuser(request):
     logout(request)
     return redirect('home')
+
+
+
+def createOffer(request,pk): 
+    if request.method == "POST":
+        title = request.POST.get('offerTitle')
+        price = request.POST.get('offerprice')
+        is_valid = request.POST.get('is_valid')
+        product = Product.objects.get(pk = pk)
+        shop = product.productOwner
+        # print(type(is_valid))
+        
+        
+        if is_valid == "on":
+            offer = Offer(product = product,offeredby = shop,offerTitle = title,offerprice = price,is_valid = True)
+            offer.save()
+        else:
+            offer = Offer(product = product,offeredby = shop,offerTitle = title,offerprice = price,is_valid = False)
+            offer.save()
+
+    return render(request,'createOffer.html',)
+
+
+def ProductOffer(request,pk):
+    product = Product.objects.get(pk = pk)
+    offers = Offer.objects.filter(product = product)
+    return render(request,'ProductOffer.html',{'offers':offers,'product':product})
